+++ src/main/java/com/consertreservation/api/seat/SeatExceptionHandler.java
@@ -1,0 +1,4 @@
+package com.consertreservation.api.seat;
+
+public class SeatExceptionHandler {
+}
