+++ src/main/java/com/consertreservation/api/usertoken/UserTokenExceptionHandler.java
@@ -1,0 +1,4 @@
+package com.consertreservation.api.usertoken;
+
+public class UserTokenExceptionHandler {
+}
