+++ src/main/java/com/consertreservation/domain/concert/repository/ConcertScheduleRepository.java
@@ -1,0 +1,8 @@
+package com.consertreservation.domain.concert.repository;
+
+import com.consertreservation.domain.concert.model.ConcertSchedule;
+
+public interface ConcertScheduleRepository {
+
+    ConcertSchedule save(ConcertSchedule concertSchedule);
+}
