+++ src/main/java/com/consertreservation/domain/concert/repository/ConcertStoreRepository.java
@@ -1,0 +1,4 @@
+package com.consertreservation.domain.concert.repository;
+
+public class ConcertStoreRepository {
+}
