+++ src/test/java/com/consertreservation/api/seat/SeatControllerTest.java
@@ -1,0 +1,4 @@
+import static org.junit.jupiter.api.Assertions.*;
+class SeatControllerTest {
+  
+}
